# improved memory management