```
benchmark                       old ns/op     new ns/op     delta
BenchmarkApplyNewObject-12      1937221       1549953       -19.99%
BenchmarkUpdateNewObject-12     2175330       2054741       -5.54%
BenchmarkRepeatedUpdate-12      761351        635385        -16.55%
BenchmarkSetToFields-12         5200          5173          -0.52%
BenchmarkFieldsToSet-12         10336         10240         -0.93%
BenchmarkYAMLToTyped-12         64579         61926         -4.11%

benchmark                       old allocs     new allocs     delta
BenchmarkApplyNewObject-12      6820           6355           -6.82%
BenchmarkUpdateNewObject-12     7901           7398           -6.37%
BenchmarkRepeatedUpdate-12      2264           2035           -10.11%
BenchmarkSetToFields-12         14             14             +0.00%
BenchmarkFieldsToSet-12         82             82             +0.00%
BenchmarkYAMLToTyped-12         152            134            -11.84%
```

(benchmarks from the main repo)

These 18 lines do as much as my entire serialization change for the bottom line, lol.