@lavalamp 2019-08-16
Adding in a few more tricks for avoiding allocations. I think at least one of the things I did shouldn't be needed, but the go escape analysis wasn't up to the task.

```
benchmark                       old ns/op     new ns/op     delta
BenchmarkApplyNewObject-12      1937221       1540317       -20.49%
BenchmarkUpdateNewObject-12     2175330       1739911       -20.02%
BenchmarkRepeatedUpdate-12      761351        533531        -29.92%
BenchmarkSetToFields-12         5200          5142          -1.12%
BenchmarkFieldsToSet-12         10336         9665          -6.49%
BenchmarkYAMLToTyped-12         64579         52677         -18.43%

benchmark                       old allocs     new allocs     delta
BenchmarkApplyNewObject-12      6820           5045           -26.03%
BenchmarkUpdateNewObject-12     7901           5839           -26.10%
BenchmarkRepeatedUpdate-12      2264           1592           -29.68%
BenchmarkSetToFields-12         14             14             +0.00%
BenchmarkFieldsToSet-12         82             82             +0.00%
BenchmarkYAMLToTyped-12         152            90             -40.79%
```

(All of these changes were done while looking at profiles, not randomly :) )