@lavalamp 2019-08-16
Effects on the local benchmarks:

```
benchmark                                              old ns/op     new ns/op     delta
BenchmarkDeducedSimple-12                              345092        343266        -0.53%
BenchmarkDeducedNested-12                              1031912       900868        -12.70%
BenchmarkDeducedNestedAcrossVersion-12                 1087190       1004928       -7.57%
BenchmarkLeafConflictAcrossVersion-12                  404455        349167        -13.67%
BenchmarkMultipleApplierRecursiveRealConversion-12     2357192       2231486       -5.33%

benchmark                                              old allocs     new allocs     delta
BenchmarkDeducedSimple-12                              1525           1264           -17.11%
BenchmarkDeducedNested-12                              4180           3341           -20.07%
BenchmarkDeducedNestedAcrossVersion-12                 4524           3602           -20.38%
BenchmarkLeafConflictAcrossVersion-12                  1605           1296           -19.25%
BenchmarkMultipleApplierRecursiveRealConversion-12     7692           6841           -11.06%

benchmark                                              old bytes     new bytes     delta
BenchmarkDeducedSimple-12                              110380        87360         -20.86%
BenchmarkDeducedNested-12                              295407        216277        -26.79%
BenchmarkDeducedNestedAcrossVersion-12                 321143        234867        -26.87%
BenchmarkLeafConflictAcrossVersion-12                  117152        89832         -23.32%
BenchmarkMultipleApplierRecursiveRealConversion-12     692692        631120        -8.89%
```